# perf: micro optimizations