The optimizations are mostly targeting decoding or encoding FIT File using `localMessageNum 0` (for both normal header or compressed timestamp header).

Here is the benchstat when decoding `testdata/from_official_sdk/activity_lowbattery.fit` (since it uses `localMessageNum 0`):

```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/decoder
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
         │  old2.txt   │              new2.txt               │
         │   sec/op    │   sec/op     vs base                │
Decode-4   13.58m ± 3%   11.97m ± 4%  -11.88% (p=0.000 n=10)

         │   old2.txt   │              new2.txt               │
         │     B/op     │     B/op      vs base               │
Decode-4   5.298Mi ± 0%   5.297Mi ± 0%  -0.03% (p=0.000 n=10)

         │  old2.txt   │              new2.txt              │
         │  allocs/op  │  allocs/op   vs base               │
Decode-4   76.22k ± 0%   76.20k ± 0%  -0.03% (p=0.000 n=10)
```

Here is the benchstat when encoding test defined FIT File:
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/encoder
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
                                     │   old.txt    │               new.txt                │
                                     │    sec/op    │    sec/op     vs base                │
Encode/normal_header_zero-4            85.75m ±  3%   87.10m ± 25%        ~ (p=0.971 n=10)
Encode/normal_header_15-4              89.67m ± 24%   80.00m ± 11%  -10.79% (p=0.007 n=10)
Encode/compressed_timestamp_header-4   79.28m ± 13%   72.77m ±  2%   -8.20% (p=0.001 n=10)
geomean                                84.79m         79.74m         -5.96%

                                     │   old.txt    │               new.txt               │
                                     │     B/op     │     B/op      vs base               │
Encode/normal_header_zero-4            3.134Mi ± 0%   3.113Mi ± 0%  -0.65% (p=0.000 n=10)
Encode/normal_header_15-4              3.049Mi ± 0%   3.057Mi ± 0%       ~ (p=0.749 n=10)
Encode/compressed_timestamp_header-4   2.342Mi ± 0%   2.327Mi ± 0%  -0.65% (p=0.000 n=10)
geomean                                2.818Mi        2.808Mi       -0.34%

                                     │   old.txt   │               new.txt                │
                                     │  allocs/op  │  allocs/op   vs base                 │
Encode/normal_header_zero-4            103.5k ± 0%   102.5k ± 0%  -0.96% (p=0.000 n=10)
Encode/normal_header_15-4              101.0k ± 0%   101.0k ± 0%       ~ (p=1.000 n=10) ¹
Encode/compressed_timestamp_header-4   103.5k ± 0%   102.5k ± 0%  -0.96% (p=0.000 n=10)
geomean                                102.7k        102.0k       -0.64%
¹ all samples are equal

```